% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_controlbar.R
\name{app_controlbar}
\alias{app_controlbar}
\alias{app_controlbar_ui}
\alias{app_controlbar_server}
\title{App Controlbar Module}
\usage{
app_controlbar_ui(id)

app_controlbar_server(id)
}
\arguments{
\item{id}{The module id}
}
\value{
\itemize{
\item \code{app_controlbar_ui()}: A bs4Dash controlbar UI element
\item \code{app_controlbar_server()}: Server logic for the controlbar
}
}
\description{
A module for creating the controlbar of the GMH Leasing Dashboard.
}
