% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_assets.R
\name{app_favicon}
\alias{app_favicon}
\title{App Favicon}
\usage{
app_favicon(path = app_sys("favicon.ico"))
}
\arguments{
\item{path}{The path to the favicon file. See details for more information
on how to set this path. Default is \code{www/images/favicons/favicon.ico},
which is specific to this package.}
}
\value{
\verb{<head>} with necessary favicon declaration
}
\description{
Add the app favicon to the app by inserting HTML tags into the app's \verb{<head>}.
}
\details{
The path to the favicon file should utilize the pre-configured \emph{resourcePath},
i.e. \code{"www/*"}, to ensure the favicon is accessible to the app.

The \code{www} prefixed path is instantiated in the \code{\link[=app_assets]{app_assets()}} function which
gets called in the \verb{<head>} of the main \code{\link[=app_ui]{app_ui()}} function.
}
