% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_footer.R
\name{app_footer}
\alias{app_footer}
\alias{app_footer_ui}
\alias{app_footer_server}
\title{App Footer Module}
\usage{
app_footer_ui(id)

app_footer_server(id)
}
\arguments{
\item{id}{The module id}
}
\value{
\itemize{
\item \code{app_footer_ui()}: A bs4Dash footer UI element
\item \code{app_footer_server()}: Server logic for the footer
}
}
\description{
A module for creating the footer of the GMH Leasing Dashboard.
}
