% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_globals.R
\name{app_globals}
\alias{app_globals}
\alias{set_app_global}
\alias{get_app_globals}
\alias{get_app_global}
\title{App Global Variables}
\usage{
set_app_global(name, value)

get_app_globals()

get_app_global(name)
}
\arguments{
\item{name}{A character string.}

\item{value}{A value.}
}
\value{
\itemize{
\item \code{set_app_global()}: Sets a global variable and invisibly returns the previous value,
if any.
\item \code{get_app_globals()}: Returns a list of all global variables.
\item \code{get_app_global()}: Returns the value of a global variable by name.
}
}
\description{
Set, get, and list global variables for the app.
}
