% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_header.R
\name{app_header}
\alias{app_header}
\alias{app_header_ui}
\alias{app_header_server}
\title{App Header Module}
\usage{
app_header_ui(id, title = "GMH Leasing Dashboard", ...)

app_header_server(id, app_globals = NULL)
}
\arguments{
\item{id}{The module id}

\item{title}{The title of the dashboard}
}
\value{
\itemize{
\item \code{app_header_ui()}: An \code{\link[htmltools:tagList]{htmltools::tagList()}} of the header UI elements
\item \code{app_header_server()}: List of reactive expressions for the header
}
}
\description{
A module for creating a custom header for the GMH Leasing Dashboard.
}
\details{
The module includes both the UI and server functions for the header:
\itemize{
\item \code{app_header_ui()}: The UI function for the header
\item \code{app_header_server()}: The server function for the header
}

The header will include the following elements:
\itemize{
\item GMH Communities logo
\item Dashboard title
\item About button
\item Refresh button
\item Contact Us dropdown menu
\item User Menu dropdown menu
}
}
