% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_assets.R
\name{app_logo}
\alias{app_logo}
\title{App Logo}
\usage{
app_logo(
  path = "public/images/logos/gmh-logo.svg",
  alt = "GMH Communities",
  style = "width: 250px; height: auto;"
)
}
\arguments{
\item{path}{Path to the logo file}

\item{alt}{Alt text for the logo}

\item{style}{CSS style for the logo}
}
\value{
HTML image tag with the provided logo
}
\description{
App Logo
}
