% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_assets.R
\name{app_meta}
\alias{app_meta}
\title{App \verb{<meta>}}
\usage{
app_meta(
  package = "gmhleasr",
  app_name = "GMH Leasing Dashboard",
  app_version = utils::packageVersion("gmhleasr"),
  description = desc::desc_get("Description")[[1]],
  url = "https://noclocks.dev",
  theme_color = "#0073AA",
  robots = "index,follow",
  generator = "RShiny",
  subject = "GMH Leasing Dashboard",
  rating = "General",
  referrer = "origin",
  csp = "default-src 'self'",
  image = "https://cdn.brandfetch.io/noclocks.dev/symbol",
  image_alt = "No Clocks, LLC Logo",
  twitter_creator = "@noclocksdev",
  twitter_card_type = "summary_large_image",
  twitter_site = "@noclocksdev",
  ...
)
}
\arguments{
\item{package}{Name of the package}

\item{app_name}{Name of the app}

\item{app_version}{Version of the app}

\item{description}{Description of the app}

\item{url}{URL of the app}

\item{theme_color}{Theme color of the app}

\item{robots}{Robots.txt policy}

\item{generator}{Generator of the app}

\item{subject}{Subject of the app}

\item{rating}{Rating of the app}

\item{referrer}{Referrer policy}

\item{csp}{Content Security Policy}

\item{image}{Image URL}

\item{image_alt}{Image alt text}

\item{twitter_creator}{Twitter creator}

\item{twitter_card_type}{Twitter card type}

\item{twitter_site}{Twitter site}

\item{...}{Additional arguments (currently unused)}
}
\value{
Raw \verb{<meta>} tags via \code{\link[metathis:meta]{metathis::meta()}} (to be passed into the
app's \verb{<head>})
}
\description{
Creates a series of \verb{<meta>} tags for the app's \verb{<head>}.
}
\seealso{
\code{\link[metathis:meta]{metathis::meta()}}
}
