% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{App Server}
\usage{
app_server(input, output, session, logging = FALSE, ...)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}.}
}
\description{
This function contains the server-side logic for the application.
}
