% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_sidebar.R
\name{app_sidebar}
\alias{app_sidebar}
\alias{app_sidebar_ui}
\alias{app_sidebar_server}
\title{App Sidebar Module}
\usage{
app_sidebar_ui(id)

app_sidebar_server(id, app_globals = NULL)

app_sidebar_ui(id)

app_sidebar_server(id, app_globals = NULL)
}
\arguments{
\item{id}{The module id}
}
\value{
\itemize{
\item \code{app_sidebar_ui()}: A bs4Dash sidebar UI element
\item \code{app_sidebar_server()}: Server logic for the sidebar
}

\itemize{
\item \code{app_sidebar_ui()}: A bs4Dash sidebar UI element
\item \code{app_sidebar_server()}: Server logic for the sidebar
}
}
\description{
A module for creating the sidebar of the GMH Leasing Dashboard.

A module for creating the sidebar of the GMH Leasing Dashboard.
}
