% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_config.R
\name{app_sys}
\alias{app_sys}
\title{App System File}
\usage{
app_sys(...)
}
\arguments{
\item{...}{Path components.}
}
\value{
Path to file in package.
}
\description{
Get the path to a file in the package.
}
\keyword{internal}
