% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{app_ui}
\alias{app_ui}
\title{App User Interface}
\usage{
app_ui(request)
}
\arguments{
\item{request}{Internal parameter for \code{{shiny}}.}
}
\value{
A \code{{shiny}} UI object.
}
\description{
The user interface for the GMH Leasing Dashboard.
}
