% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cache.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear Cache}
\usage{
clear_cache(
  cache_dir = getOption("gmh.cache_dir", default = rappdirs::user_cache_dir("gmhleasr"))
)
}
\arguments{
\item{cache_dir}{The directory path to the cache.}
}
\description{
Clear the cache and remove all cached files.
}
\details{
This function will remove all files and directories under the specified cache directory.
}
