% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui_contact.R
\name{create_contact_menu}
\alias{create_contact_menu}
\title{Create Contact Menu}
\usage{
create_contact_menu(contacts = .app_contacts())
}
\arguments{
\item{contacts}{list of contacts (name, role, company, phone, email, url, image)}
}
\value{
menu
}
\description{
Creates a dropdown menu specific for contacts
}
