% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_config.R
\name{create_entrata_config}
\alias{create_entrata_config}
\title{Create Entrata API Configuration}
\usage{
create_entrata_config(..., config_path = "config.yml")
}
\arguments{
\item{...}{Named arguments to pass to APIConfig$new()}

\item{config_path}{Path to a YAML configuration file. Default is "config.yml" in the current directory.}
}
\value{
An APIConfig object
}
\description{
Create a new Entrata API configuration object. By default, it looks for a config.yml file
in the current working directory. If found, it uses the configuration from this file.
Otherwise, it uses the provided arguments or falls back to default values.
}
