% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_request.R
\name{derive_req_body}
\alias{derive_req_body}
\title{Derive Entrata API Request Body}
\usage{
derive_req_body(method, method_version, method_params)
}
\arguments{
\item{method}{The Entrata API method to use.}

\item{method_version}{The version of the API method to use.}

\item{method_params}{A named list of parameters to include in the request body.}
}
\value{
A list representing the request body.
}
\description{
Derives the request body for an Entrata API request.
}
