% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_text.R
\name{display_phone_number}
\alias{display_phone_number}
\title{Displau Phone Number}
\usage{
display_phone_number(phone)
}
\arguments{
\item{phone}{String representing a phone number.}
}
\value{
Formatted string of the phone number.
}
\description{
Display phone number in a formatted string.
}
\examples{
display_phone_number("5555555555")
}
