% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_config.R
\name{entrata_config}
\alias{entrata_config}
\alias{APIConfig}
\title{Entrata API Configuration}
\description{
Configure settings and options related to the Entrata API.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{username}}{API username}

\item{\code{password}}{API password}

\item{\code{base_url}}{Base URL for API requests}

\item{\code{user_agent}}{User agent string for API requests}

\item{\code{rate_limit}}{Rate limit for API requests}

\item{\code{cache_dir}}{Directory for caching API responses}

\item{\code{cache_ttl}}{Time-to-live for cached responses (in seconds)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-APIConfig-new}{\code{APIConfig$new()}}
\item \href{#method-APIConfig-print}{\code{APIConfig$print()}}
\item \href{#method-APIConfig-clone}{\code{APIConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-APIConfig-new"></a>}}
\if{latex}{\out{\hypertarget{method-APIConfig-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new APIConfig object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{APIConfig$new(
  username = NULL,
  password = NULL,
  base_url = NULL,
  user_agent = NULL,
  rate_limit = NULL,
  cache_dir = NULL,
  cache_ttl = 3600
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{username}}{API username}

\item{\code{password}}{API password}

\item{\code{base_url}}{Base URL for API requests}

\item{\code{user_agent}}{User agent string for API requests}

\item{\code{rate_limit}}{Rate limit for API requests}

\item{\code{cache_dir}}{Directory for caching API responses}

\item{\code{cache_ttl}}{Time-to-live for cached responses (in seconds)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-APIConfig-print"></a>}}
\if{latex}{\out{\hypertarget{method-APIConfig-print}{}}}
\subsection{Method \code{print()}}{
Print a summary of the configuration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{APIConfig$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-APIConfig-clone"></a>}}
\if{latex}{\out{\hypertarget{method-APIConfig-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{APIConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
