% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_internal.R
\name{entrata_internal}
\alias{entrata_internal}
\alias{validate_entrata_endpoint_method}
\alias{validate_entrata_method_params}
\alias{get_default_method}
\title{Entrata API Internal Helpers}
\usage{
validate_entrata_endpoint_method(endpoint, method)

validate_entrata_method_params(endpoint, method, method_params)

get_default_method(endpoint)
}
\arguments{
\item{endpoint}{Entrata API Endpoint to validate.}

\item{method}{Entrata API Method to validate.}

\item{method_params}{List of parameters to validate.}
}
\description{
Internal helper functions for the Entrata API request functions.
\itemize{
\item \code{validate_entrata_endpoint_method()}: Validate Entrata API Endpoint and Method
\item \code{validate_entrata_method_params()}: Validate Entrata API Method Parameters
\item \code{get_default_method()}: Get Default Method by Endpoint
}
}
\seealso{
\href{https://docs.entrata.com/api/v1/documentation/}{Entrata API Documentation}
}
\keyword{internal}
