% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_internal.R
\name{entrata_request_validation}
\alias{entrata_request_validation}
\alias{validate_entrata_request_endpoint}
\alias{validate_entrata_request_endpoint_method}
\alias{validate_entrata_request_method_params}
\alias{validate_entrata_request}
\title{Entrata API Request Validations}
\usage{
validate_entrata_request_endpoint(
  endpoint,
  arg = rlang::caller_arg(endpoint),
  call = rlang::caller_env()
)

validate_entrata_request_endpoint_method(
  endpoint,
  method,
  arg_endpoint = rlang::caller_arg(endpoint),
  arg_method = rlang::caller_arg(method),
  call = rlang::caller_env()
)

validate_entrata_request_method_params(
  endpoint,
  method,
  method_params,
  arg_endpoint = rlang::caller_arg(endpoint),
  arg_method = rlang::caller_arg(method),
  arg_method_params = rlang::caller_arg(method_params),
  call = rlang::caller_env()
)

validate_entrata_request(
  endpoint,
  method,
  method_params,
  arg_endpoint = rlang::caller_arg(endpoint),
  arg_method = rlang::caller_arg(method),
  arg_method_params = rlang::caller_arg(method_params),
  call = rlang::caller_env(),
  ...
)
}
\arguments{
\item{endpoint}{Entrata API Endpoint to validate. Must be one of the available
Entrata API endpoints.}

\item{arg, arg_endpoint, arg_method, arg_method_params}{Arguments to use for error messages.}

\item{call}{Environment to use for error messages. Defaults to the calling environment.}

\item{method}{Entrata API Endpoint Method to validate. Should be one of the
available methods for the specified endpoint.}

\item{method_params}{List of method parameters to validate. Should be a named list
corresponding the the specific endpoint/method's available and required parameters.}

\item{...}{Additional arguments to pass to the underlying validation functions.}
}
\value{
All validation functions are used for their side effects have no
return values, except for \code{validate_entrata_request()} which will
return \code{TRUE} if validations pass.
}
\description{
Helper functions for validating Entrata API requests.
\itemize{
\item \code{validate_entrata_request_endpoint()}: Validate Entrata API Endpoint
\item \code{validate_entrata_request_endpoint_method()}: Validate Entrata API Endpoint's Method
\item \code{validate_entrata_request_method_params()}: Validate Entrata API Endpoint Method's Parameters
\item \code{validate_entrata_request()}: Validate Entrata API Request by performing all validation checks
}
}
\seealso{
\href{https://docs.entrata.com/api/v1/documentation/}{Entrata API Documentation}
}
