% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_resp_status.R
\name{entrata_resp_status}
\alias{entrata_resp_status}
\alias{entrata_resp_status_desc}
\alias{entrata_resp_is_error}
\alias{entrata_resp_abort}
\alias{entrata_resp_check_status}
\title{Entrata API Response Status Helpers}
\usage{
entrata_resp_status(resp)

entrata_resp_status_desc(resp)

entrata_resp_is_error(resp)

entrata_resp_abort(resp, req, info = NULL, call = rlang::caller_env())

entrata_resp_check_status(resp, info = NULL, error_call = rlang::caller_env())
}
\arguments{
\item{resp}{\code{\link[httr2:response]{httr2::response()}} object}

\item{info}{Additional information to include in the error message.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{error_call}{The calling environment for the error. Defaults to the
caller environment.}
}
\value{
\itemize{
\item \code{entrata_resp_status()} - The status code of the response.
\item \code{entrata_resp_status_desc()} - The status description of the response.
\item \code{entrata_resp_is_error()} - Logical value indicating whether the response is an error.
\item \code{entrata_resp_check_status()} - Invisibly returns the response if the status is not an error.
\item \code{entrata_resp_abort()} - Aborts the response if the status is an error.
}
}
\description{
Helper functions for working with the Entrata API response status codes,
descriptions, and errors.
\itemize{
\item \code{entrata_resp_status()} retrieves the integer \code{HTTP} status code from an
Entrata API response.
\item \code{entrata_resp_status_desc()} retrieves a brief textual description of the
\code{HTTP} status code from an Entrata API response.
\item \code{entrata_resp_is_error()} determines if the response is an error, i.e.
it returns \code{TRUE} if the status code is greater than or equal to \code{400}.
\item \code{entrata_resp_check_status()} checks the status of an Entrata API response
by turning HTTP/Entrata specific errors into R errors.
\item \code{entrata_resp_abort()} aborts the response if the status is an error.
}

Note that these functions are mostly for internal use but are exported
for clarity and consistency.
}
\details{
These functions are used to help with the processing of Entrata API
responses.

The Entrata API is unique in that all endpoints utilize \code{POST} \code{HTTP} methods
and the API will always return a \code{200} status code, regardless of the success
or failure of the request.

This makes the usage of the default \code{httr2} functions like \code{\link[httr2:resp_status]{httr2::resp_check_status()}}
or \code{\link[httr2:resp_status]{httr2::resp_is_error()}} useless as they rely on the \code{HTTP} status code to
determine if a request was successful or not.

To mitigate this, we must inspect the response body to determine if the request
was successful or not. If the response body contains an error message, we can
assume that the request was not successful.
}
\section{Functions}{

\itemize{
\item \code{entrata_resp_status()} - Extract the status code from an Entrata API response.
\item \code{entrata_resp_status_desc()} - Extract the status description from an Entrata API response.
\item \code{entrata_resp_is_error()} - Determine if the response is an error.
\item \code{entrata_resp_check_status()} - Check the status of an Entrata API response.
\item \code{entrata_resp_abort()} - Abort the requested response using the response object.
}
}

\seealso{
\code{\link[httr2:resp_status]{httr2::resp_status()}}
}
