% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_validation.R
\name{entrata_validation}
\alias{entrata_validation}
\alias{validate_entrata_request_endpoint}
\alias{validate_entrata_request_endpoint_method}
\alias{validate_entrata_request_method_params}
\alias{validate_entrata_request}
\alias{validate_entrata_report_name}
\title{Entrata API Request Validations}
\usage{
validate_entrata_request_endpoint(endpoint)

validate_entrata_request_endpoint_method(endpoint, method)

validate_entrata_request_method_params(endpoint, method, method_params)

validate_entrata_request(endpoint, method, method_params, ...)

validate_entrata_report_name(report_name, ...)
}
\arguments{
\item{endpoint}{Entrata API Endpoint to validate.}

\item{method}{Entrata API Endpoint Method to validate.}

\item{method_params}{List of method parameters to validate.}

\item{...}{Additional arguments to pass to the underlying validation functions.}

\item{report_name}{Report name to validate.}
}
\value{
All validation functions are used for their side effects and have no
return values, except for \code{validate_entrata_request()} which will
return \code{TRUE} if validations pass.
}
\description{
Helper functions for validating Entrata API requests.
}
\seealso{
\href{https://docs.entrata.com/api/v1/documentation/}{Entrata API Documentation}
}
