% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_config.R
\name{get_app_config}
\alias{get_app_config}
\title{Get App Config}
\usage{
get_app_config(
  value,
  config = Sys.getenv("APP_CONFIG_ACTIVE", Sys.getenv("R_CONFIG_ACTIVE", "default")),
  use_parent = TRUE,
  file = Sys.getenv("APP_CONFIG_FILE", Sys.getenv("R_CONFIG_FILE",
    app_sys("config/config.yml")))
)
}
\arguments{
\item{value}{Name of value (\code{NULL} to read all values)}

\item{config}{Name of configuration to read from. Defaults to
the value of the \code{R_CONFIG_ACTIVE} environment variable
("default" if the variable does not exist).}

\item{use_parent}{\code{TRUE} to scan parent directories for
configuration files if the specified config file isn't found.}

\item{file}{Configuration file to read from (defaults to
\code{"config.yml"}). If the file isn't found at the location
specified then parent directories are searched for a file
of the same name.}
}
\value{
The value of the configuration key.
}
\description{
Retrieve configuration values from a the app's \code{config.yml} \code{YAML} file.
}
\details{
The only difference in this function and \code{\link[config:get]{config::get()}} is that this function
sets the default value of the \code{file} argument to the app's \code{config.yml} file
and attempts to set the \code{config} argument to the value of the \code{APP_CONFIG_ACTIVE}.
}
