% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_reports.R
\name{get_latest_report_version}
\alias{get_latest_report_version}
\alias{mem_get_latest_report_version}
\title{Get Latest Report Version}
\usage{
get_latest_report_version(report_name)

mem_get_latest_report_version(report_name)
}
\arguments{
\item{report_name}{A character string representing the name of the report.}
}
\value{
A character string representing the latest version of the report.
}
\description{
This function retrieves the latest version of a specified
report from Entrata.
}
\section{Functions}{
\itemize{
\item \code{mem_get_latest_report_version()}: Memoized Get Latest Report Version

}}
