% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_properties.R
\name{get_properties}
\alias{get_properties}
\title{Get Entrata Properties}
\usage{
get_properties(
  api,
  property_ids = NULL,
  property_lookup_codes = NULL,
  show_all_status = FALSE,
  ...
)
}
\arguments{
\item{api}{An instance of the \code{EntrataAPI} class.}

\item{property_ids}{A vector of property IDs to retrieve.}

\item{property_lookup_codes}{A vector of property lookup codes to retrieve.}

\item{show_all_status}{Logical. Whether to return all properties regardless of status.}

\item{...}{Additional parameters to pass to the underlying API request.}
}
\value{
A data frame with the retrieved property information.
}
\description{
Retrieves properties from the Entrata API.
}
