% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_reports_pre_lease.R
\name{get_property_ids_filter_param}
\alias{get_property_ids_filter_param}
\title{Get Property IDs Filter Parameter}
\usage{
get_property_ids_filter_param()
}
\value{
A list of character strings representing property IDs.
}
\description{
This function retrieves the list of property IDs to be used
as a filter parameter in Entrata API requests.
}
