% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_inputs.R
\name{input_picker}
\alias{input_picker}
\title{Picker Input}
\usage{
input_picker(
  id,
  label,
  choices,
  selected = choices,
  multiple = TRUE,
  count_threshold = 3,
  choice_options = NULL,
  ...
)
}
\arguments{
\item{id}{Input ID for the picker}

\item{label}{Label for the picker}

\item{choices}{Choices for the picker}

\item{selected}{Selected choices}

\item{multiple}{Are multiple selections allowed?}

\item{count_threshold}{Threshold for displaying count of selected items,
default is \code{3}.}

\item{choice_options}{Passed through to \code{\link[shinyWidgets:pickerInput]{shinyWidgets::pickerInput()}}'s \code{choicesOpt}
argument.}

\item{...}{Additional arguments passed to \code{\link[shinyWidgets:pickerInput]{shinyWidgets::pickerInput()}}.}
}
\value{
A select input control that can be added to the app's UI.
}
\description{
\code{input_picker()} is a function to create a generic picker (select) input for
a Shiny App. It wraps the \code{\link[shinyWidgets:pickerInput]{shinyWidgets::pickerInput()}} function.
}
\seealso{
\code{\link[shinyWidgets:pickerInput]{shinyWidgets::pickerInput()}}
}
