% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_boolean_string}
\alias{is_boolean_string}
\title{Is Boolean String}
\usage{
is_boolean_string(str)
}
\arguments{
\item{str}{Character string to check. Typically, with the Entrata API, "boolean"
values are represented as quoted integers (\code{"0"} or \code{"1"}) representing
\code{FALSE} and \code{TRUE}, respectively.}
}
\value{
\code{TRUE} if the string is a boolean string, \code{FALSE} otherwise.
}
\description{
Checks if provided string represents a boolean (used by Entrata API).
}
\examples{
is_boolean_string("0")
}
