% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_validation.R
\name{is_integer_string}
\alias{is_integer_string}
\title{Check if a string represents an integer}
\usage{
is_integer_string(str)
}
\arguments{
\item{str}{String to check}
}
\value{
Logical indicating if the string represents an integer
}
\description{
Check if a string represents an integer
}
\keyword{internal}
