% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_validation.R
\name{is_integer_string_multi}
\alias{is_integer_string_multi}
\title{Check if a string represents multiple integers}
\usage{
is_integer_string_multi(str)
}
\arguments{
\item{str}{String to check}
}
\value{
Logical indicating if the string represents multiple integers
}
\description{
Check if a string represents multiple integers
}
\keyword{internal}
