% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_config.R
\name{load_config}
\alias{load_config}
\title{Load Configuration from YAML}
\usage{
load_config(path)
}
\arguments{
\item{path}{Path to the YAML configuration file}
}
\value{
A list of configuration parameters
}
\description{
Load Entrata API configuration from a YAML file. It expects a structure with a "default"
section containing an "entrata" subsection with the configuration parameters.
}
\keyword{internal}
