% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reactivity.R
\name{make_reactive_trigger}
\alias{make_reactive_trigger}
\title{Make Reactive Trigger}
\usage{
make_reactive_trigger(dev = FALSE, label = "")
}
\arguments{
\item{dev}{logical; if TRUE, will print messages to console}

\item{label}{character; label for the trigger}
}
\value{
list with two functions: \code{depend()} and \code{trigger()}
}
\description{
Create a 'reactive trigger' object (not a proper R class) with two methods:
\enumerate{
\item \code{depend()} : must be written in a code chunk to execute on triggering
\item \code{trigger()}: when executed, trigger the object (and all the "depending"
code chunks)
}

This function is freely re-used from Dean Attali's work:
See \href{https://github.com/daattali/advanced-shiny/tree/master/reactive-trigger}{here}.
}
\examples{

# create a reactive trigger
trigger <- make_reactive_trigger(dev = TRUE, label = "my_trigger")
}
