% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_header_elements.R
\name{mod_header_elements_server}
\alias{mod_header_elements_server}
\title{Header Elements Server}
\usage{
mod_header_elements_server(
  id,
  parent_session = NULL,
  paths = list(help = fs::path_package("gmhrleasr", "assets/content/help.Rmd"), about =
    fs::path_package("gmhrleasr", "assets/content/about.Rmd"))
)
}
\arguments{
\item{parent_session}{shiny session of the parent environment}

\item{paths}{named list of file paths}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\value{
server
}
\description{
Server module for header elements.
}
