% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_header_elements.R
\name{mod_header_elements_ui}
\alias{mod_header_elements_ui}
\title{Header Elements UI}
\usage{
mod_header_elements_ui(
  id,
  include_elements = list(help = TRUE, about = TRUE, refresh = TRUE, contact = TRUE, user
    = TRUE),
  contacts = NULL
)
}
\arguments{
\item{id}{namespace ID}

\item{include_elements}{named list of logicals for element inclusion}

\item{contacts}{list of contact information}
}
\value{
\code{htmltools::tagList()}
}
\description{
UI module for header elements.
}
