% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_properties.R
\name{mod_properties}
\alias{mod_properties}
\alias{mod_properties_ui}
\alias{mod_properties_server}
\title{Properties Module}
\usage{
mod_properties_ui(id)

mod_properties_server(id, ...)
}
\arguments{
\item{id}{The module id}

\item{...}{Additional parameters}

\item{input, output, session}{Internal parameters for {shiny}.}
}
\value{
\itemize{
\item \code{mod_properties_ui()}: An \code{\link[htmltools:tagList]{htmltools::tagList()}} of the properties UI elements
\item \code{mod_properties_server()}: List of reactive expressions for the properties
}
}
\description{
A module for displaying the GMH Communities Properties data.
}
\details{
The module includes both the UI and server functions for the properties:
\itemize{
\item \code{mod_properties_ui()}: The UI function for the properties
\item \code{mod_properties_server()}: The server function for the properties
}

The properties module will include the following elements:
\itemize{
\item Properties Table
\item Properties Map
\item Properties Chart
\item Properties Info
}
}
