% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sidebar.R
\name{mod_sidebar}
\alias{mod_sidebar}
\alias{mod_sidebar_ui}
\alias{mod_sidebar_server}
\title{Dashboard Sidebar Module}
\usage{
mod_sidebar_ui(id)

mod_sidebar_server(id)
}
\arguments{
\item{id}{The namespace ID for the module.}
}
\value{
\itemize{
\item \code{mod_sidebar_ui()} returns the UI for the dashboard sidebar.
\item \code{mod_sidebar_server()} returns the server logic for the dashboard sidebar.
}
}
\description{
\itemize{
\item \code{mod_sidebar_ui()} creates the UI for the dashboard sidebar.
\item \code{mod_sidebar_server()} creates the server logic for the dashboard sidebar.
}
}
