% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_leases.R
\name{parse_entrata_leases}
\alias{parse_entrata_leases}
\alias{parse_entrata_lease_customers}
\alias{parse_entrata_lease_intervals}
\alias{parse_entrata_lease_scheduled_charges}
\alias{parse_entrata_lease_unit_spaces}
\title{Parse Response for Entrata Leases}
\usage{
parse_entrata_leases(res)

parse_entrata_lease_customers(res_content)

parse_entrata_lease_intervals(res_content)

parse_entrata_lease_scheduled_charges(res_content)

parse_entrata_lease_unit_spaces(res_content)
}
\arguments{
\item{res}{The \code{\link[httr2:response]{httr2::response()}} object from the Entrata API}

\item{res_content}{Response content to parse}
}
\value{
Parsed Response Body Content as a tibble with leases data.

Parsed Response Body Content as a tibble with lease customer data.

Parsed Response Body Content as a tibble with lease interval data.

Parsed Response Body Content as a tibble with lease scheduled charges data.

Parsed Response Body Content as a tibble with lease unit spaces data.
}
\description{
This function parses the response from the Entrata API's "getLeases" method.
}
\details{
The core function is \code{parse_entrata_leases()}, which parses the response
from the Entrata API's "getLeases" method called via \code{\link[=entrata_leases]{entrata_leases()}}.

\code{parse_entrata_leases()} calls the following functions to parse the response:
\itemize{
\item \code{parse_entrata_lease_customers()}: Parse Entrata lease customers
\item \code{parse_entrata_lease_intervals()}: Parse Entrata lease intervals
\item \code{parse_entrata_lease_scheduled_charges()}: Parse Entrata lease scheduled charges
\item \code{parse_entrata_lease_unit_spaces()}: Parse Entrata lease unit spaces
}
}
\section{Functions}{
\itemize{
\item \code{parse_entrata_lease_customers()}: Parse Entrata Lease Customers

\item \code{parse_entrata_lease_intervals()}: Parse Entrata Lease Intervals

\item \code{parse_entrata_lease_scheduled_charges()}: Parse Entrata Lease Scheduled Charges

\item \code{parse_entrata_lease_unit_spaces()}: Parse Entrata Lease Unit Spaces

}}
\seealso{
\code{\link[=entrata_leases]{entrata_leases()}}
}
