% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_inputs.R
\name{picker_entrata_reports}
\alias{picker_entrata_reports}
\title{Picker Input for Entrata Reports}
\usage{
picker_entrata_reports(
  id,
  label = "Select Report",
  selected = "pre_lease",
  ...
)
}
\arguments{
\item{id}{id}

\item{label}{label}

\item{selected}{selected}

\item{...}{dots}
}
\value{
a
}
\description{
Picker Input for Entrata Reports
}
