% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_reports.R
\name{prep_pre_lease_report_params}
\alias{prep_pre_lease_report_params}
\title{Prepare Pre-Lease Report Parameters}
\usage{
prep_pre_lease_report_params(
  latest_report_version = mem_get_latest_report_version("pre_lease"),
  property_group_ids = mem_get_property_ids_filter_param(),
  period_start_date = "09/01/2024",
  period_type = "date",
  summarize_by = "unit_type",
  group_by = "unit_type",
  consider_pre_leased_on = "33",
  charge_code_detail = 1,
  space_options = "do_not_show",
  additional_units_shown = "available",
  combine_unit_spaces_with_same_lease = 0,
  consolidate_by = "no_consolidation",
  arrange_by_property = 0,
  subtotals = list("summary", "details"),
  yoy = 1,
  ...
)
}
\arguments{
\item{latest_report_version}{A character string representing the latest
version of the pre-lease report. Defaults to fetching the latest version.}

\item{property_group_ids}{A list of property group IDs to include in the report.}

\item{period_start_date}{The start date of the reporting period. Defaults to "09/01/2024".}

\item{period_type}{The type of period for the report. Defaults to "date".}

\item{summarize_by}{The level at which to summarize the report. Defaults to "unit_type".}

\item{group_by}{The level at which to group the report. Defaults to "unit_type".}

\item{consider_pre_leased_on}{A string representing the pre-leased on date. Defaults to "33".}

\item{charge_code_detail}{An integer for charge code details. Defaults to 1.}

\item{space_options}{A string representing space options. Defaults to "do_not_show".}

\item{additional_units_shown}{A string representing additional units shown. Defaults to "available".}

\item{combine_unit_spaces_with_same_lease}{An integer indicating whether to combine unit spaces. Defaults to 0.}

\item{consolidate_by}{A string indicating consolidation options. Defaults to "no_consolidation".}

\item{arrange_by_property}{An integer indicating whether to arrange by property. Defaults to 0.}

\item{subtotals}{A list of subtotals to include in the report. Defaults to c("summary", "details").}

\item{yoy}{An integer indicating whether to include year-over-year data. Defaults to 1.}

\item{...}{Additional parameters for the report.}
}
\value{
A list of parameters for the pre-lease report.
}
\description{
This function prepares the parameters required for generating a
pre-lease report in Entrata.
}
