% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_internal.R
\name{request_error}
\alias{request_error}
\alias{res_err_body}
\alias{res_is_err}
\title{Request Error Helpers}
\usage{
res_err_body(resp)

res_is_err(resp)
}
\arguments{
\item{resp}{\code{\link[httr2:response]{httr2::response()}} object}
}
\value{
\itemize{
\item \code{res_err_body()} - Error message as a string.
\item \code{res_is_err()} - Logical value indicating whether the response is an error.
\item \code{res_err_retry()} - Logical value indicating whether the request is retryable.
}
}
\description{
Helper functions for \code{\link[httr2:req_error]{httr2::req_error()}}. These functions are used in
conjunction with the \code{req_error()} function to provide additional information
about the error that occurred during the request.
}
\details{
The Entrata API is unique in that it always returns a \code{200} status code,
regardless of the success or failure of the request. To determine if a request
was successful or not, we must inspect the response body. If the response body
contains an error message, we can assume that the request was not successful.
}
\section{Functions}{

\itemize{
\item \code{res_err_body()} - Extracts the error code and message from the response
body. Used for the \code{body} argument in \code{\link[httr2:req_error]{httr2::req_error()}}.
\item \code{res_is_err()} - Determines if the response is an error. This is used for the
\code{is_error} argument in \code{\link[httr2:req_error]{httr2::req_error()}}.
\item \code{res_err_retry()} - Determines if the request should be retried based off
the response's error code.
}
}

\seealso{
\code{\link[httr2:req_error]{httr2::req_error()}}
}
