% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_run.R
\name{run_with_opts}
\alias{run_with_opts}
\title{Run with Options}
\usage{
run_with_opts(
  app,
  opts,
  showcase = FALSE,
  maintenance_page = maintenance_page,
  print = TRUE
)
}
\arguments{
\item{app}{Shiny App object.}

\item{opts}{A list of options to pass to the Shiny App.}

\item{maintenance_page}{A Shiny App object to display when the app is in maintenance mode.}

\item{print}{A logical value to print the app. Defaults to \code{TRUE}.}
}
\value{
A Shiny App object.
}
\description{
Run a Shiny App with extra, custom options.
}
\keyword{internal}
