% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_config.R
\name{save_config}
\alias{save_config}
\title{Save Configuration to YAML}
\usage{
save_config(config, path)
}
\arguments{
\item{config}{An APIConfig object}

\item{path}{Path to save the YAML configuration file}
}
\value{
Invisible NULL
}
\description{
Save Entrata API configuration to a YAML file.
}
