% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_text.R
\name{strip_phone_number}
\alias{strip_phone_number}
\title{Strip Phone Number}
\usage{
strip_phone_number(phone)
}
\arguments{
\item{phone}{String representing a phone number.}
}
\value{
Formatted string of the phone number.
}
\description{
Strip phone number of all non-numeric characters and add country code.
Used to format phone numbers properly for use with the \verb{tel://} \code{HTML}
attribute's protocol.
}
\examples{
strip_phone_number("(555) 555-5555")
strip_phone_number("555-555-5555")
strip_phone_number("555.555.5555")
}
