% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_text.R
\name{trim_ws}
\alias{trim_ws}
\alias{trim_ws.character}
\alias{trim_ws.data.frame}
\alias{trim_ws.list}
\title{Trim Whitespace}
\usage{
trim_ws(str, ...)

\method{trim_ws}{character}(x, ...)

\method{trim_ws}{data.frame}(x, character_only = TRUE, ...)

\method{trim_ws}{list}(x, character_only = TRUE, ...)
}
\arguments{
\item{str}{Character vector, or a data frame.}

\item{...}{Currently not used.}
}
\value{
Character vector, where trailing and leading white spaces are removed.
}
\description{
\code{trim_ws()} is an efficient function to trim leading and trailing
white space from character vectors or strings.
}
\examples{
trim_ws("  no space!  ")
}
