% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_html.R
\name{utils_html}
\alias{utils_html}
\alias{centered}
\alias{help_mark}
\alias{help_label}
\title{HTML Utilities}
\usage{
centered(...)

help_mark(help_txt)

help_label(label_text, help_text)
}
\arguments{
\item{...}{HTML elements}

\item{help_txt}{Help text}

\item{label_text}{Label text}
}
\value{
\itemize{
\item \code{centered()}: Centered HTML element
\item \code{help_mark()}: Help icon with hover text
\item \code{help_label()}: Label with help
}
}
\description{
\itemize{
\item \code{centered()}: Centered HTML element
\item \code{help_mark()}: Help icon with hover text
\item \code{help_label()}: Label with help icon
}
}
