% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_schemas.R
\name{utils_schema}
\alias{utils_schema}
\alias{write_json_schema}
\alias{yaml_to_json}
\alias{yaml_to_list}
\title{Schema (JSON/YAML) Utilities}
\usage{
write_json_schema(schema, path)

yaml_to_json(yaml)

yaml_to_list(yaml)
}
\arguments{
\item{schema}{JSON schema}

\item{path}{File path}
}
\value{
\itemize{
\item \code{write_json_schema()}: Invisible file path
\item \code{yaml_to_json()}: JSON string
\item \code{yaml_to_list()}: List
}
}
\description{
\itemize{
\item \code{write_json_schema()}: Write JSON schema to file
\item \code{yaml_to_json()}: Convert YAML to JSON
\item \code{yaml_to_list()}: Convert YAML to list
}
}
