% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_config.R
\name{validate_entrata_config}
\alias{validate_entrata_config}
\title{Validate Entrata API Configuration}
\usage{
validate_entrata_config(cfg)
}
\arguments{
\item{cfg}{Either a path to a YAML configuration file or a list object containing the configurations.}
}
\value{
The original configuration object, invisibly.
}
\description{
Validate proper configuration setup for Entrata API credentials and URLs.
}
