% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_reports.R
\name{validate_entrata_report_name}
\alias{validate_entrata_report_name}
\title{Validate Entrata Report Name}
\usage{
validate_entrata_report_name(report_name)
}
\arguments{
\item{report_name}{A character string representing the report name to validate.}
}
\value{
NULL. Throws an error if the report name is invalid.
}
\description{
This function checks if a provided report name exists in the
list of available Entrata reports. If the report name is not valid,
an error is thrown with suggestions for valid report names.
}
