% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_css.R
\name{verify_css}
\alias{verify_css}
\title{Verify CSS}
\usage{
verify_css(selector, css)
}
\arguments{
\item{selector}{The CSS selector for the element to verify}

\item{css}{A named list of CSS properties and their expected values}
}
\value{
A logical value indicating whether the CSS properties match the expected values
}
\description{
Verify that the CSS properties of a given element match the expected values.
}
\examples{
\dontrun{
verify_css("#my-element", list("color" = "red", "font-size" = "16px"))
}

}
