% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_leases.R
\name{entrata_leases}
\alias{entrata_leases}
\title{Entrata Leases}
\usage{
entrata_leases(
  property_ids,
  application_id = NULL,
  customer_id = NULL,
  lease_status_type_ids = NULL,
  lease_ids = NULL,
  scheduled_ar_code_ids = NULL,
  unit_number = NULL,
  building_name = NULL,
  move_in_date_from = NULL,
  move_in_date_to = NULL,
  lease_expiring_date_from = NULL,
  lease_expiring_date_to = NULL,
  move_out_date_from = NULL,
  move_out_date_to = NULL,
  include_other_income_leases = NULL,
  resident_friendly_mode = NULL,
  include_lease_history = NULL,
  include_ar_transactions = NULL,
  pagination_page_number = 1,
  pagination_page_size = 500,
  include_pagination_links = FALSE,
  ...
)
}
\arguments{
\item{property_ids}{Required. Integer vector of property IDs}

\item{application_id}{Optional. Integer value of the application ID}

\item{customer_id}{Optional. Integer value of the customer ID}

\item{lease_status_type_ids}{Optional. Integer vector of lease status type IDs}

\item{lease_ids}{Optional. Integer vector of lease IDs}

\item{scheduled_ar_code_ids}{Optional. Integer vector of scheduled AR code IDs}

\item{unit_number}{Optional. Character value of the unit number}

\item{building_name}{Optional. Character value of the building name}

\item{move_in_date_from}{Optional. Date value of the move-in date from}

\item{move_in_date_to}{Optional. Date value of the move-in date to}

\item{lease_expiring_date_from}{Optional. Date value of the lease expiring from}

\item{lease_expiring_date_to}{Optional. Date value of the lease expiring to}

\item{move_out_date_from}{Optional. Date value of the move-out date from}

\item{move_out_date_to}{Optional. Date value of the move-out date to}

\item{include_other_income_leases}{Optional. Logical value to include other income leases}

\item{resident_friendly_mode}{Optional. Logical value to include resident friendly mode}

\item{include_lease_history}{Optional. Logical value to include lease history information}

\item{include_ar_transactions}{Optional. Logical value to include AR transactions}

\item{pagination_page_number}{Pagination page number. Default is 1}

\item{pagination_page_size}{Number of items per page. Default is 500}

\item{include_pagination_links}{Logical value to include pagination links in the response. Default is FALSE}

\item{...}{Additional parameters to pass to the request}
}
\value{
Parsed Response Body Content as a tibble with leases data.
}
\description{
This function retrieves lease information from the Entrata API.
}
\seealso{
\code{\link[=parse_entrata_leases]{parse_entrata_leases()}}
}
