% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_reports.R
\name{get_latest_report_version}
\alias{get_latest_report_version}
\title{Get Latest Report Version from Entrata}
\usage{
get_latest_report_version(report_name)
}
\arguments{
\item{report_name}{A character string representing the name of the report.}
}
\value{
A character string representing the latest version of the report.
}
\description{
This function retrieves the latest version of a specified report from Entrata.
It's useful when you need to ensure you're working with the most up-to-date
version of a particular report.
}
\examples{
\dontrun{
# Get the latest version of the "pre_lease" report
latest_version <- get_latest_report_version("pre_lease")
print(latest_version)
}

}
\seealso{
\code{\link{get_entrata_reports_list}} for getting a list of all available reports
\code{\link{get_entrata_report_info}} for getting detailed information about a specific report
}
