% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmhleasr-package.R
\docType{package}
\name{gmhleasr-package}
\alias{gmhleasr}
\alias{gmhleasr-package}
\title{gmhleasr: Shiny Dashboard for GMH Communities Leasing Data}
\description{
Shiny Dashboard for GMH Communities Leasing Data.
}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})

Authors:
\itemize{
  \item Patrick Howard \email{patrick.howard@noclocks.dev} [reviewer]
}

Other contributors:
\itemize{
  \item No Clocks, LLC \email{team@noclocks.dev} [copyright holder, funder]
}

}
\keyword{internal}
