% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmhleasr-package.R
\docType{package}
\name{gmhleasr-package}
\alias{gmhleasr}
\alias{gmhleasr-package}
\title{GMH Leasing R Package - \code{gmhleasr}}
\description{
The \code{gmhleasr} package provides a comprehensive set of tools for interacting with the Entrata API.
It includes functions for managing API configurations, making API requests, and parsing responses.
The package is designed to simplify the process of retrieving and analyzing data from Entrata,
particularly focusing on leasing and property management information.
}
\details{
\emph{R package housing a shiny application and data processing logic related to
the leasing activities at GMH Communities.}
}
\section{Main Features}{

\itemize{
\item API Configuration Management: Easy setup and management of Entrata API credentials and settings.
\item Request Handling: Robust functions for making API requests with built-in error handling and retrying capabilities.
\item Response Parsing: Tools for parsing and cleaning API responses into usable R data structures.
\item Caching: Efficient caching mechanisms to improve performance and reduce API calls.
\item Data Retrieval: Specialized functions for retrieving property, lease, and report data from Entrata.
}
}

\section{Key Functions}{

\itemize{
\item \code{\link{create_entrata_config}}: Create and manage Entrata API configurations.
\item \code{\link{entrata}}: Make requests to the Entrata API.
\item \code{\link{entrata_properties}}: Retrieve property information from Entrata.
\item \code{\link{entrata_leases}}: Fetch lease data from Entrata.
\item \code{\link{get_entrata_reports_list}}: Get a list of available reports from Entrata.
\item \code{\link{get_entrata_report_info}}: Retrieve detailed information about specific Entrata reports.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/noclocks/gmhleasr}
  \item \url{https://docs.noclocks.dev/gmhleasr/}
  \item Report bugs at \url{https://github.com/noclocks/gmhleasr/issues}
}

}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})

Authors:
\itemize{
  \item Patrick Howard \email{patrick.howard@noclocks.dev} [reviewer]
}

Other contributors:
\itemize{
  \item No Clocks, LLC \email{team@noclocks.dev} [copyright holder, funder]
}

}
